adds workflow struct and TestPreconditions function
* renames package cmdb to infocmdb (go convention)
* moves workflow related types and functions to separate workflow.go file
* renames Workflow to WorkflowFunc
* adds workflow struct with Config field
* moves RunWorkflow function to type Workflow and shortens it to Run.
* adds TestPreconditions function to Workflow for easy table tests.

New workflow example:
```go
func main() {
	w := infocmdb.NewWorkflow()
	w.Run(workflow)
}

func workflow(params infocmdb.WorkflowParams, cmdb *infocmdb.Client) (err error) {
	...
}
```

New workflow preconditions test:
```go
func TestPreconditions(t *testing.T) {
	w := infocmdb.NewWorkflow()
	w.TestPreconditions(t, infocmdb.Preconditions{
		{infocmdb.TYPE_ATTRIBUTE, "attribute_1"},
		{infocmdb.TYPE_ATTRIBUTE, "attribute_2"},
		{infocmdb.TYPE_RELATION, "relation_1"},
	})
}
```
